requirements.txt
aiohttp>=3.7.4
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==5.2.0
python-dotenv==0.19.2
typing-extensions==3.10.0.2
APScheduler~=3.8.1
yarl==1.7.2
